+++ packages/at_client/lib/src/service/sync_service_impl.dart
@@ -10,7 +10,6 @@ import 'package:at_client/src/response/default_response_parser.dart';
 import 'package:at_client/src/response/json_utils.dart';
 import 'package:at_client/src/service/notification_service_impl.dart';
 import 'package:at_client/src/service/sync/sync_request.dart';
-import 'package:at_client/src/service/sync_service.dart';
 import 'package:at_client/src/util/logger_util.dart';
 import 'package:at_client/src/util/network_util.dart';
 import 'package:at_client/src/util/sync_util.dart';
@@ -624,21 +623,30 @@ class SyncServiceImpl implements SyncService, AtSignChangeListener {
       List<CommitEntry> uncommittedEntries) async {
     var batchRequests = <BatchRequest>[];
     var batchId = 1;
+    List<CommitEntry> removeUncommittedEntriesList = [];
     for (var entry in uncommittedEntries) {
       String command;
-      // If someone updates cached keys locally, they should never be synced to the cloud
+      // The update on a cached key is prevented. The logic in "validatePutRequest"
+      // throws exception if a user tries to update a cached key.
+      // The below check is for the older data. The cached keys that are updated
+      // before the "validatePutRequest" is in-place.
       // However if they want to delete a cached key, they should be allowed to
       if (entry.atKey!.startsWith('cached:') &&
           entry.operation != CommitOp.DELETE) {
         _logger.finer(
             '${entry.atKey} is skipped. cached keys will not be synced to cloud secondary');
+        removeUncommittedEntriesList.add(entry);
         continue;
       }
+      // For CommitOp.Update, _getCommand fetches the data from the local keystore to sync to the server.
+      // When getCommand is called for an entry where key is created/updated and then deleted,
+      // a KeyNotFoundException will be thrown because the data does not exist in the keystore.
       try {
         command = await _getCommand(entry);
       } on KeyNotFoundException {
         _logger.severe(
             '${entry.atKey} is not found in keystore. Skipping to entry to sync');
+        removeUncommittedEntriesList.add(entry);
         continue;
       }
       command = VerbUtil.replaceNewline(command);
@@ -647,6 +655,18 @@ class SyncServiceImpl implements SyncService, AtSignChangeListener {
       batchRequests.add(batchRequest);
       batchId++;
     }
+    // The commit-id's in the batch response are updated to the appropriate commit-entry
+    // in the uncommitted entries by iterating the uncommitted entries list.
+    // If an entry is skipped in the batch request, then size of batch response
+    // will be less than the size of uncommitted entries and so the commit-id gets
+    // updated against the wrong uncommitted entry.
+    // So, remove the commit entry from the uncommitted entries list.
+    for (CommitEntry commitEntry in removeUncommittedEntriesList) {
+      uncommittedEntries.remove(commitEntry);
+      // Removing the entry from the commit log keystore to prevent stale entries
+      await syncUtil.removeCommitEntry(commitEntry.key, currentAtSign);
+    }
+    removeUncommittedEntriesList.clear();
     return batchRequests;
   }